Reusing getWindowOptions and simplified import of process-directories
